+++ include/xrpl/core/Job.h
@@ -1,10 +1,9 @@
 #ifndef XRPL_CORE_JOB_H_INCLUDED
 #define XRPL_CORE_JOB_H_INCLUDED
 
-#include <xrpld/core/ClosureCounter.h>
-#include <xrpld/core/LoadMonitor.h>
-
 #include <xrpl/basics/CountedObject.h>
+#include <xrpl/core/ClosureCounter.h>
+#include <xrpl/core/LoadMonitor.h>
 
 #include <functional>
