+++ include/xrpl/core/JobQueue.h
@@ -1,12 +1,11 @@
 #ifndef XRPL_CORE_JOBQUEUE_H_INCLUDED
 #define XRPL_CORE_JOBQUEUE_H_INCLUDED
 
-#include <xrpld/core/ClosureCounter.h>
-#include <xrpld/core/JobTypeData.h>
-#include <xrpld/core/JobTypes.h>
-#include <xrpld/core/detail/Workers.h>
-
 #include <xrpl/basics/LocalValue.h>
+#include <xrpl/core/ClosureCounter.h>
+#include <xrpl/core/JobTypeData.h>
+#include <xrpl/core/JobTypes.h>
+#include <xrpl/core/detail/Workers.h>
 #include <xrpl/json/json_value.h>
 
 #include <boost/coroutine/all.hpp>
@@ -385,7 +384,7 @@ private:
 
 }  // namespace ripple
 
-#include <xrpld/core/Coro.ipp>
+#include <xrpl/core/Coro.ipp>
 
 namespace ripple {
