+++ include/xrpl/core/JobTypeData.h
@@ -1,10 +1,9 @@
 #ifndef XRPL_CORE_JOBTYPEDATA_H_INCLUDED
 #define XRPL_CORE_JOBTYPEDATA_H_INCLUDED
 
-#include <xrpld/core/JobTypeInfo.h>
-
 #include <xrpl/basics/Log.h>
 #include <xrpl/beast/insight/Collector.h>
+#include <xrpl/core/JobTypeInfo.h>
 
 namespace ripple {
