+++ include/xrpl/core/JobTypeInfo.h
@@ -1,7 +1,7 @@
 #ifndef XRPL_CORE_JOBTYPEINFO_H_INCLUDED
 #define XRPL_CORE_JOBTYPEINFO_H_INCLUDED
 
-#include <xrpld/core/Job.h>
+#include <xrpl/core/Job.h>
 
 namespace ripple {
