+++ include/xrpl/core/LoadMonitor.h
@@ -1,10 +1,9 @@
 #ifndef XRPL_CORE_LOADMONITOR_H_INCLUDED
 #define XRPL_CORE_LOADMONITOR_H_INCLUDED
 
-#include <xrpld/core/LoadEvent.h>
-
 #include <xrpl/basics/UptimeClock.h>
 #include <xrpl/beast/utility/Journal.h>
+#include <xrpl/core/LoadEvent.h>
 
 #include <chrono>
 #include <mutex>
