+++ include/xrpl/core/detail/Workers.h
@@ -1,9 +1,8 @@
 #ifndef XRPL_CORE_WORKERS_H_INCLUDED
 #define XRPL_CORE_WORKERS_H_INCLUDED
 
-#include <xrpld/core/detail/semaphore.h>
-
 #include <xrpl/beast/core/LockFreeStack.h>
+#include <xrpl/core/detail/semaphore.h>
 
 #include <atomic>
 #include <condition_variable>
