+++ src/libxrpl/core/detail/Job.cpp
@@ -1,6 +1,5 @@
-#include <xrpld/core/Job.h>
-
 #include <xrpl/beast/core/CurrentThreadName.h>
+#include <xrpl/core/Job.h>
 
 namespace ripple {
