+++ src/libxrpl/core/detail/JobQueue.cpp
@@ -1,7 +1,7 @@
-#include <xrpld/core/JobQueue.h>
 #include <xrpld/perflog/PerfLog.h>
 
 #include <xrpl/basics/contract.h>
+#include <xrpl/core/JobQueue.h>
 
 #include <mutex>
