+++ src/libxrpl/core/detail/Workers.cpp
@@ -1,8 +1,8 @@
-#include <xrpld/core/detail/Workers.h>
 #include <xrpld/perflog/PerfLog.h>
 
 #include <xrpl/beast/core/CurrentThreadName.h>
 #include <xrpl/beast/utility/instrumentation.h>
+#include <xrpl/core/detail/Workers.h>
 
 namespace ripple {
