+++ src/test/app/Path_test.cpp
@@ -4,11 +4,11 @@
 #include <test/jtx/envconfig.h>
 #include <test/jtx/permissioned_dex.h>
 
-#include <xrpld/core/JobQueue.h>
 #include <xrpld/rpc/RPCHandler.h>
 #include <xrpld/rpc/detail/Tuning.h>
 
 #include <xrpl/beast/unit_test.h>
+#include <xrpl/core/JobQueue.h>
 #include <xrpl/json/json_reader.h>
 #include <xrpl/protocol/ApiVersion.h>
 #include <xrpl/protocol/STParsedJSON.h>
