+++ src/test/app/Transaction_ordering_test.cpp
@@ -1,6 +1,6 @@
 #include <test/jtx.h>
 
-#include <xrpld/core/JobQueue.h>
+#include <xrpl/core/JobQueue.h>
 
 namespace ripple {
 namespace test {
