+++ src/test/core/ClosureCounter_test.cpp
@@ -1,8 +1,7 @@
 #include <test/jtx/Env.h>
 
-#include <xrpld/core/ClosureCounter.h>
-
 #include <xrpl/beast/unit_test.h>
+#include <xrpl/core/ClosureCounter.h>
 
 #include <atomic>
 #include <chrono>
