+++ src/test/core/JobQueue_test.cpp
@@ -1,8 +1,7 @@
 #include <test/jtx/Env.h>
 
-#include <xrpld/core/JobQueue.h>
-
 #include <xrpl/beast/unit_test.h>
+#include <xrpl/core/JobQueue.h>
 
 namespace ripple {
 namespace test {
