+++ src/test/core/Workers_test.cpp
@@ -1,7 +1,7 @@
-#include <xrpld/core/detail/Workers.h>
 #include <xrpld/perflog/PerfLog.h>
 
 #include <xrpl/beast/unit_test.h>
+#include <xrpl/core/detail/Workers.h>
 #include <xrpl/json/json_value.h>
 
 #include <chrono>
