+++ src/test/rpc/RobustTransaction_test.cpp
@@ -1,9 +1,8 @@
 #include <test/jtx.h>
 #include <test/jtx/WSClient.h>
 
-#include <xrpld/core/JobQueue.h>
-
 #include <xrpl/beast/unit_test.h>
+#include <xrpl/core/JobQueue.h>
 #include <xrpl/protocol/jss.h>
 
 namespace ripple {
