+++ src/xrpld/app/ledger/OrderBookDB.cpp
@@ -4,9 +4,9 @@
 #include <xrpld/app/misc/AMMUtils.h>
 #include <xrpld/app/misc/NetworkOPs.h>
 #include <xrpld/core/Config.h>
-#include <xrpld/core/JobQueue.h>
 
 #include <xrpl/basics/Log.h>
+#include <xrpl/core/JobQueue.h>
 #include <xrpl/protocol/Indexes.h>
 
 namespace ripple {
