+++ src/xrpld/app/ledger/detail/TimeoutCounter.h
@@ -2,9 +2,9 @@
 #define XRPL_APP_LEDGER_TIMEOUTCOUNTER_H_INCLUDED
 
 #include <xrpld/app/main/Application.h>
-#include <xrpld/core/Job.h>
 
 #include <xrpl/beast/utility/Journal.h>
+#include <xrpl/core/Job.h>
 
 #include <boost/asio/basic_waitable_timer.hpp>
