+++ src/xrpld/app/main/NodeStoreScheduler.h
@@ -1,8 +1,7 @@
 #ifndef XRPL_APP_MAIN_NODESTORESCHEDULER_H_INCLUDED
 #define XRPL_APP_MAIN_NODESTORESCHEDULER_H_INCLUDED
 
-#include <xrpld/core/JobQueue.h>
-
+#include <xrpl/core/JobQueue.h>
 #include <xrpl/nodestore/Scheduler.h>
 
 namespace ripple {
