+++ src/xrpld/app/misc/NetworkOPs.h
@@ -3,9 +3,9 @@
 
 #include <xrpld/app/consensus/RCLCxPeerPos.h>
 #include <xrpld/app/ledger/Ledger.h>
-#include <xrpld/core/JobQueue.h>
 #include <xrpld/rpc/InfoSub.h>
 
+#include <xrpl/core/JobQueue.h>
 #include <xrpl/ledger/ReadView.h>
 #include <xrpl/protocol/STValidation.h>
 #include <xrpl/protocol/messages.h>
