+++ src/xrpld/app/paths/Pathfinder.cpp
@@ -4,10 +4,10 @@
 #include <xrpld/app/paths/RippleCalc.h>
 #include <xrpld/app/paths/RippleLineCache.h>
 #include <xrpld/app/paths/detail/PathfinderUtils.h>
-#include <xrpld/core/JobQueue.h>
 
 #include <xrpl/basics/Log.h>
 #include <xrpl/basics/join.h>
+#include <xrpl/core/JobQueue.h>
 #include <xrpl/json/to_string.h>
 #include <xrpl/ledger/PaymentSandbox.h>
