+++ src/xrpld/app/paths/Pathfinder.h
@@ -3,9 +3,9 @@
 
 #include <xrpld/app/ledger/Ledger.h>
 #include <xrpld/app/paths/RippleLineCache.h>
-#include <xrpld/core/LoadEvent.h>
 
 #include <xrpl/basics/CountedObject.h>
+#include <xrpl/core/LoadEvent.h>
 #include <xrpl/protocol/STAmount.h>
 #include <xrpl/protocol/STPathSet.h>
