+++ src/xrpld/overlay/detail/OverlayImpl.h
@@ -2,7 +2,6 @@
 #define XRPL_OVERLAY_OVERLAYIMPL_H_INCLUDED
 
 #include <xrpld/app/main/Application.h>
-#include <xrpld/core/Job.h>
 #include <xrpld/overlay/Message.h>
 #include <xrpld/overlay/Overlay.h>
 #include <xrpld/overlay/Slot.h>
@@ -16,6 +15,7 @@
 #include <xrpl/basics/UnorderedContainers.h>
 #include <xrpl/basics/chrono.h>
 #include <xrpl/beast/utility/instrumentation.h>
+#include <xrpl/core/Job.h>
 #include <xrpl/resource/ResourceManager.h>
 #include <xrpl/server/Handoff.h>
