+++ src/xrpld/overlay/detail/PeerSet.cpp
@@ -1,7 +1,8 @@
 #include <xrpld/app/main/Application.h>
-#include <xrpld/core/JobQueue.h>
 #include <xrpld/overlay/Overlay.h>
 #include <xrpld/overlay/PeerSet.h>
+
+#include <xrpl/core/JobQueue.h>
 
 namespace ripple {
