+++ src/xrpld/perflog/PerfLog.h
@@ -2,8 +2,8 @@
 #define XRPL_BASICS_PERFLOG_H
 
 #include <xrpld/core/Config.h>
-#include <xrpld/core/JobTypes.h>
 
+#include <xrpl/core/JobTypes.h>
 #include <xrpl/json/json_value.h>
 
 #include <boost/filesystem.hpp>
