+++ src/xrpld/perflog/detail/PerfLogImp.cpp
@@ -1,9 +1,9 @@
-#include <xrpld/core/JobTypes.h>
 #include <xrpld/perflog/detail/PerfLogImp.h>
 
 #include <xrpl/basics/BasicConfig.h>
 #include <xrpl/beast/core/CurrentThreadName.h>
 #include <xrpl/beast/utility/Journal.h>
+#include <xrpl/core/JobTypes.h>
 #include <xrpl/json/json_writer.h>
 
 #include <atomic>
