+++ src/xrpld/rpc/ServerHandler.h
@@ -3,9 +3,9 @@
 
 #include <xrpld/app/main/Application.h>
 #include <xrpld/app/main/CollectorManager.h>
-#include <xrpld/core/JobQueue.h>
 #include <xrpld/rpc/detail/WSInfoSub.h>
 
+#include <xrpl/core/JobQueue.h>
 #include <xrpl/json/Output.h>
 #include <xrpl/server/Server.h>
 #include <xrpl/server/Session.h>
