+++ src/xrpld/rpc/detail/LegacyPathFind.cpp
@@ -1,9 +1,10 @@
 #include <xrpld/app/main/Application.h>
 #include <xrpld/app/misc/LoadFeeTrack.h>
-#include <xrpld/core/Job.h>
-#include <xrpld/core/JobQueue.h>
 #include <xrpld/rpc/detail/LegacyPathFind.h>
 #include <xrpld/rpc/detail/Tuning.h>
+
+#include <xrpl/core/Job.h>
+#include <xrpl/core/JobQueue.h>
 
 namespace ripple {
 namespace RPC {
